requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
async-timeout=4.0.3=py311hca03da5_0
blinker=1.6.2=py311hca03da5_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.7.2=hecd8cb5_0
click=8.1.7=py311hca03da5_0
deprecated=1.2.13=py311hca03da5_0
flask=3.0.3=py311hca03da5_0
itsdangerous=2.2.0=py311hca03da5_0
jinja2=3.1.4=py311hca03da5_0
libffi=3.4.4=hecd8cb5_1
markupsafe=2.1.3=py311h80987f9_0
ncurses=6.4=hcec6c5f_0
openssl=3.0.14=h80987f9_0
packaging=24.1=py311hca03da5_0
pip=24.2=py311hecd8cb5_0
python=3.11.9=hf27a42d_0
readline=8.2=hca72f7f_0
redis-py=4.3.4=py311hca03da5_0
setuptools=72.1.0=py311hecd8cb5_0
sqlite=3.45.3=h6c40b1e_0
tk=8.6.14=h4d00af3_0
tzdata=2024a=h04d1e81_0
werkzeug=3.0.3=py311hca03da5_0
wheel=0.43.0=py311hecd8cb5_0
wrapt=1.14.1=py311h80987f9_0
xz=5.4.6=h6c40b1e_1
zlib=1.2.13=h4b97444_1
